requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
pydantic>=1.8.0
loguru==0.5.3
google-cloud-storage==2.10.0
pandas==2.1.1